+++ stats.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstdint>
+#include <initializer_list>
 #include <optional>
 #include <string>
 
@@ -13,5 +14,9 @@ struct StatsValue {
 };
 
 StatsValue<int32_t> stats_value(int32_t value, bool required = true);
+
+StatsValue<int32_t> merge_stats_values(
+    std::initializer_list<StatsValue<int32_t>> values
+);
 
 enum class MapStage { pre, main, post };